Move the Utils class into the utils package
