+++ app/src/main/java/fr/free/nrw/commons/WelcomePagerAdapter.java
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import androidx.viewpager.widget.PagerAdapter;
+import fr.free.nrw.commons.utils.Utils;
 
 public class WelcomePagerAdapter extends PagerAdapter {
     private static final int[] PAGE_LAYOUTS = new int[]{
