+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsFragment.kt
@@ -30,7 +30,7 @@ import androidx.work.WorkManager
 import fr.free.nrw.commons.MapController.NearbyPlacesInfo
 import fr.free.nrw.commons.Media
 import fr.free.nrw.commons.R
-import fr.free.nrw.commons.Utils
+import fr.free.nrw.commons.utils.Utils
 import fr.free.nrw.commons.auth.SessionManager
 import fr.free.nrw.commons.campaigns.CampaignView
 import fr.free.nrw.commons.campaigns.CampaignsPresenter
